refactor: move transform graphql variables method somewhere else
